feat: add honey command and separate db
