+++ README.md
@@ -1,7 +1,7 @@
 # flang-bot
 
-이 저장소는 간단한 디스코드 봇 예제입니다. 봇은 `/인사`와 `/가입` 두 가지 명령을 제공합니다.
-`/인사`는 인사 메시지를 반환하고, `/가입`은 명령을 실행한 사용자의 정보를 SQLite 데이터베이스에 저장합니다. 저장되는 정보에는 아이디와 사용자 이름은 물론 프로필 이미지 URL과 명령을 실행한 서버에서의 닉네임도 포함됩니다.
+이 저장소는 간단한 디스코드 봇 예제입니다. 봇은 `/인사`, `/가입`, `/꿀단지` 세 가지 명령을 제공합니다.
+`/인사`는 인사 메시지를 반환하고, `/가입`은 명령을 실행한 사용자의 정보를 SQLite 데이터베이스에 저장합니다. 저장되는 정보에는 아이디와 사용자 이름은 물론 프로필 이미지 URL과 명령을 실행한 서버에서의 닉네임도 포함됩니다. `/꿀단지` 명령을 사용하면 저장된 자신의 정보를 확인할 수 있습니다.
 
 ## 사용 방법
